pyproject.toml
[tool.poetry]
name = "trabalho"
version = "0.1.0"
description = ""
authors = ["Igor Cardoso <igorgcardoso7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
numpy = "^1.26.2"
glob2 = "^0.7"
soundfile = "^0.12.1"
scipy = "^1.11.4"
configparser = "^6.0.0"
datasets = { extras = ["audio"], version = "^2.15.0" }
safetensors = "^0.4.1"
torchmetrics = "^1.2.1"
pillow = "^10.1.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
torchinfo = "^1.8.0"
utils = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
